requirements.txt
bleach==3.3.0
build==0.5.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
docutils==0.17.1
idna==2.10
importlib-metadata==4.5.0
jeepney==0.6.0
keyring==23.0.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pep517==0.10.0
pkginfo==1.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-cricket-scraper==0.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
toml==0.10.2
tqdm==4.61.1
twine==3.4.1
urllib3==1.26.5
webencodings==0.5.1
zipp==3.4.1
